Refactor autoload/.../substitute.vim
Create new functions for adding missing delimiters, adding the old
search pattern if no search pattern is given and wrapping the search
pattern in \%V. Change PerformSubstitution into PrepareAndSubstitute and
make use of the previously named functions.

Let repetitions of a substition also use PrepareAndSubstitute such that
their search pattern is retrieved as well as wrapped in \%V to avoid
bugs where & and s: operate linewise and not just in the visually
selected area.

Additionally, create an operator-function to repeat the last
substitution without the last used flags.

Moreover, do not throw some vim-error when the used delimiter is ", \ or
| - use echoerr.