Fix bug that breaks operator-usage with motions
A previous commit fixed the substitution of the rightmost character in a
selection but introduced another bug as well: Using the
substitution-operator with a motion does not substitute anything and
enters linewise visual-mode.  Manually setting the ']-mark one step to
the right is an alternate solution to the original problem and does not
break the operator-usage with motions.