Do not call inputsave() and inputrestore()
Calling these functions removes the possibility to give input for the
substitute operator in recordings, feedkeys(), mappings and (as a
result) tests. The desired behavior is, however, that keys are used as
input once the operator is called and as long as <CR> is not pressed.

Moreover, remove superfluous echo command.