Handle input-strings consisting only of delimiters
The keepempty argument in the split()-call is used to avoid that E648 is
thrown when trying to obtain the zero-element of the resulting list (as
the list would be empty for an input-string consisting only of
delimiters).

Do not update the /-register when the search pattern is empty as the old
search pattern is used in that case.